Extract tooltip view
